+++ djangocms_picture/cms_plugins.py
@@ -1,12 +1,4 @@
 # -*- coding: utf-8 -*-
-
-from __future__ import unicode_literals
-
-try:
-    import urlparse
-except ImportError:
-    from urllib import parse as urlparse
-
 from django.conf import settings
 from django.utils.translation import ugettext_lazy as _
 
@@ -14,33 +6,73 @@ from cms.plugin_base import CMSPluginBase
 from cms.plugin_pool import plugin_pool
 
 from .models import Picture
+from .forms import PictureForm
+
+
+# enable nesting of plugins inside the picture plugin
+PICTURE_NESTING = getattr(settings, 'DJANGOCMS_PICTURE_NESTING', False)
 
 
 class PicturePlugin(CMSPluginBase):
     model = Picture
-    name = _("Picture")
-    render_template = "cms/plugins/picture.html"
+    form = PictureForm
+    name = _('Image')
+    allow_children = PICTURE_NESTING
     text_enabled = True
 
-    def render(self, context, instance, placeholder):
-        if instance.url:
-            link = instance.url
-        elif instance.page_link:
-            link = instance.page_link.get_absolute_url()
-        else:
-            link = ""
-        context.update({
-            'picture': instance,
-            'link': link,
-            'placeholder': placeholder
+    fieldsets = [
+        (None, {
+            'fields': (
+                'picture',
+                'external_picture',
+            )
+        }),
+        (_('Advanced settings'), {
+            'classes': ('collapse',),
+            'fields': (
+                'template',
+                ('width', 'height'),
+                'alignment',
+                'caption_text',
+                'attributes',
+            )
+        }),
+        (_('Link settings'), {
+            'classes': ('collapse',),
+            'fields': (
+                ('link_url', 'link_page'),
+                'link_target',
+                'link_attributes',
+            )
+        }),
+        (_('Cropping settings'), {
+            'classes': ('collapse',),
+            'fields': (
+                ('use_automatic_scaling', 'use_no_cropping'),
+                ('use_crop', 'use_upscale'),
+                'thumbnail_options',
+            )
         })
-        return context
+    ]
 
-    def icon_src(self, instance):
-        if getattr(settings, 'PICTURE_FULL_IMAGE_AS_ICON', False):
-            return instance.image.url
-        else:
-            return urlparse.urljoin(
-                settings.STATIC_URL, "cms/img/icons/plugins/image.png")
+    def get_render_template(self, context, instance, placeholder):
+        return 'djangocms_picture/{}/picture.html'.format(instance.template)
+
+    def render(self, context, instance, placeholder):
+        if instance.alignment:
+            classes = 'align-{} '.format(instance.alignment)
+            classes += instance.attributes.get('class', '')
+            # Set the class attribute to include the alignment html class
+            # This is done to leverage the attributes_str property
+            instance.attributes['class'] = classes
+        # assign link to a context variable to be performant
+        context['picture_link'] = instance.get_link()
+        context['picture_size'] = instance.get_size(
+            width=context.get('width'),
+            height=context.get('height'),
+        )
+
+        return super(PicturePlugin, self).render(context, instance, placeholder)
+
 
 plugin_pool.register_plugin(PicturePlugin)