+++ djangocms_picture/models.py
@@ -1,83 +1,286 @@
 # -*- coding: utf-8 -*-
-
-from __future__ import unicode_literals
-
-import os
-
-from django.core.exceptions import ValidationError
+"""
+Enables the user to add an "Image" plugin that displays an image
+using the HTML <img> tag.
+"""
 from django.db import models
-from django.utils.translation import ugettext_lazy as _
+from django.conf import settings
+from django.core.exceptions import ValidationError
+from django.utils.encoding import python_2_unicode_compatible
+from django.utils.translation import ugettext, ugettext_lazy as _
 
-from cms.models import CMSPlugin, Page
-try:
-    from cms.models import get_plugin_media_path
-except ImportError:
-    def get_plugin_media_path(instance, filename):
-        """
-        See cms.models.pluginmodel.get_plugin_media_path on django CMS 3.0.4+
-        for information
-        """
-        return instance.get_media_path(filename)
-from cms.utils.compat.dj import python_2_unicode_compatible
+from cms.models import CMSPlugin
+from cms.models.fields import PageField
+
+from djangocms_attributes_field.fields import AttributesField
+
+from filer.models import ThumbnailOption
+from filer.fields.image import FilerImageField
+
+
+# add setting for picture alignment, renders a class or inline styles
+# depending on your template setup
+PICTURE_ALIGNMENT = getattr(
+    settings,
+    'DJANGOCMS_PICTURE_ALIGN',
+    (
+        ('left', _('Align left')),
+        ('right', _('Align right')),
+        ('left', _('Align center')),
+    )
+)
+
+# use golden ration as default (https://en.wikipedia.org/wiki/Golden_ratio)
+PICTURE_RATIO = getattr(settings, 'DJANGOCMS_PICTURE_RATIO', 1.6180)
+
+LINK_TARGET = (
+    ('_blank', _('Open in new window.')),
+    ('_self', _('Open in same window.')),
+    ('_parent', _('Delegate to parent.')),
+    ('_top', _('Delegate to top.')),
+)
+
+# Add additional choices through the ``settings.py``.
+def get_templates():
+    choices = [
+        ('default', _('Default')),
+    ]
+    choices += getattr(
+        settings,
+        'DJANGOCMS_PICTURE_TEMPLATES',
+        [],
+    )
+    return choices
 
 
 @python_2_unicode_compatible
 class Picture(CMSPlugin):
     """
-    A Picture with or without a link.
+    Renders an image with the option of adding a link
     """
-    LEFT = "left"
-    RIGHT = "right"
-    CENTER = "center"
-    FLOAT_CHOICES = ((LEFT, _("left")),
-                     (RIGHT, _("right")),
-                     (CENTER, _("center")),
-                     )
+    TEMPLATE_CHOICES = [
+        ('default', _('Default')),
+    ]
 
-    image = models.ImageField(_("image"), upload_to=get_plugin_media_path)
-    url = models.CharField(
-        _("link"), max_length=255, blank=True, null=True,
-        help_text=_("If present, clicking on image will take user to link."))
+    template = models.CharField(
+        verbose_name=_('Template'),
+        choices=get_templates(),
+        default=TEMPLATE_CHOICES[0][0],
+        max_length=255,
+    )
+    picture = FilerImageField(
+        verbose_name=_('Image'),
+        blank=True,
+        null=True,
+        on_delete=models.SET_NULL,
+        related_name='+',
+    )
+    external_picture = models.URLField(
+        verbose_name=_('External image'),
+        blank=True,
+        max_length=255,
+        help_text=_('If provided, overrides the embedded image. '
+            'Certain options such as cropping are not applicable for external images.')
+    )
+    width = models.PositiveIntegerField(
+        verbose_name=_('Width'),
+        blank=True,
+        null=True,
+        help_text=_('The image width as number in pixels. '
+            'Example: "720" and not "720px".'),
+    )
+    height = models.PositiveIntegerField(
+        verbose_name=_('Height'),
+        blank=True,
+        null=True,
+        help_text=_('The image height as number in pixels. '
+            'Example: "720" and not "720px".'),
+    )
+    alignment = models.CharField(
+        verbose_name=_('Alignment'),
+        choices=PICTURE_ALIGNMENT,
+        blank=True,
+        max_length=255,
+        help_text=_('Aligns the image to the selected option.'),
+    )
+    caption_text = models.TextField(
+        verbose_name=_('Caption text'),
+        blank=True,
+        help_text=_('Provide a description, attribution, copyright or other information.')
+    )
+    attributes = AttributesField(
+        verbose_name=_('Attributes'),
+        blank=True,
+        excluded_keys=['src', 'width', 'height'],
+    )
+    # link models
+    link_url = models.URLField(
+        verbose_name=_('External URL'),
+        blank=True,
+        max_length=255,
+        help_text=_('Wraps the image in a link to an external URL.'),
+    )
+    link_page = PageField(
+        verbose_name=_('Internal URL'),
+        blank=True,
+        null=True,
+        on_delete=models.SET_NULL,
+        help_text=_('Wraps the image in a link to an internal (page) URL.'),
+    )
+    link_target = models.CharField(
+        verbose_name=_('Link target'),
+        choices=LINK_TARGET,
+        blank=True,
+        max_length=255,
+    )
+    link_attributes = AttributesField(
+        verbose_name=_('Link attributes'),
+        blank=True,
+        excluded_keys=['href', 'target'],
+    )
+    # cropping models
+    # active per default
+    use_automatic_scaling = models.BooleanField(
+        verbose_name=_('Automatic scaling'),
+        blank=True,
+        default=True,
+        help_text=_('Uses the placeholder dimenstions to automatically calculate the size.'),
+    )
+    # ignores all other cropping options
+    # throws validation error if other cropping options are selected
+    use_no_cropping = models.BooleanField(
+        verbose_name=_('Use original image.'),
+        blank=True,
+        default=False,
+        help_text=_('Outputs the raw image without cropping.'),
+    )
+    # upscale and crop work together
+    # throws validation error if other cropping options are selected
+    use_crop = models.BooleanField(
+        verbose_name=_('Crop image'),
+        blank=True,
+        default=False,
+        help_text=_('Crops the image according to the given thumbnail settings in the template.'),
+    )
+    use_upscale = models.BooleanField(
+        verbose_name=_('Upscale image'),
+        blank=True,
+        default=False,
+        help_text=_('Upscales the image to the size of the thumbnail settings in the template.')
+    )
+    # overrides all other options
+    # throws validation error if other cropping options are selected
+    thumbnail_options = models.ForeignKey(
+        ThumbnailOption,
+        verbose_name=_('Thumbnail options'),
+        blank=True,
+        null=True,
+        help_text=_('Overrides width, height, crop and upscale with the provided preset.'),
+    )
 
-    page_link = models.ForeignKey(
-        Page, verbose_name=_("page"), null=True,
-        limit_choices_to={'publisher_is_draft': True}, blank=True,
-        help_text=_("If present, clicking on image will take user to "
-                    "specified page."))
-
-    alt = models.CharField(
-        _("alternate text"), max_length=255, blank=True, null=True,
-        help_text=_("Specifies an alternate text for an image, if the image"
-                    "cannot be displayed.<br />Is also used by search engines"
-                    "to classify the image."))
-
-    longdesc = models.CharField(
-        _("long description"), max_length=255, blank=True, null=True,
-        help_text=_("When user hovers above picture, this text will appear "
-                    "in a popup."))
-
-    float = models.CharField(
-        _("side"), max_length=10, blank=True, null=True, choices=FLOAT_CHOICES,
-        help_text=_("Move image left, right or center."))
-
-    width = models.IntegerField(_("width"), blank=True, null=True,
-                                help_text=_("Pixel"))
-    height = models.IntegerField(_("height"), blank=True, null=True,
-                                 help_text=_("Pixel"))
+    # Add an app namespace to related_name to avoid field name clashes
+    # with any other plugins that have a field with the same name as the
+    # lowercase of the class name of this model.
+    # https://github.com/divio/django-cms/issues/5030
+    cmsplugin_ptr = models.OneToOneField(
+        CMSPlugin,
+        related_name='%(app_label)s_%(class)s',
+        parent_link=True,
+    )
 
     def __str__(self):
-        if self.alt:
-            return self.alt[:40]
-        elif self.image:
-            # added if, because it raised attribute error when file wasn't
-            # defined.
-            try:
-                return u"%s" % os.path.basename(self.image.name)
-            except AttributeError:
-                pass
-        return u"<empty>"
+        if self.picture and self.picture.label:
+            return self.picture.label
+        return str(self.pk)
+
+    def get_short_description(self):
+        if self.picture and self.picture.label:
+            return self.picture.label
+        if self.link_url:
+            return self.link_url
+        return ugettext('<file is missing>')
+
+    def copy_relations(self, oldinstance):
+        # Because we have a ForeignKey, it's required to copy over
+        # the reference from the instance to the new plugin.
+        self.picture = oldinstance.picture
+
+    def get_size(self, width=None, height=None):
+        crop = self.use_crop
+        upscale = self.use_upscale
+        # use field thumbnail settings
+        if self.thumbnail_options:
+            width = self.thumbnail_options.width
+            height = self.thumbnail_options.height
+            crop = self.thumbnail_options.crop
+            upscale = self.thumbnail_options.upscale
+        elif not self.use_automatic_scaling:
+            width = self.width
+            height = self.height
+
+        # calculate height when not given according to the
+        # golden ratio or fallback to the picture size
+        if not height and width:
+            height = int(width / PICTURE_RATIO)
+        elif not width and height:
+            width = int(height * PICTURE_RATIO)
+        elif not width and not height:
+            width = self.picture.width
+            height = self.picture.height
+
+        options = {
+            'size': (width, height),
+            'crop': crop,
+            'upscale': upscale,
+        }
+        return options
+
+    def get_link(self):
+        if self.link_url:
+            return self.link_url
+        if self.link_page_id:
+            return self.link_page.get_absolute_url(language=self.language)
+        return False
 
     def clean(self):
-        if self.url and self.page_link:
+        # there can be only one link type
+        if self.link_page_id and self.link_page:
             raise ValidationError(
-                _("You can enter a Link or a Page, but not both."))
+                ugettext('You have defined an external and internal link. '
+                         'Only one option is allowed.')
+            )
+
+        # you shall only set one image kind
+        if not self.picture and not self.external_picture:
+            raise ValidationError(
+                ugettext('You need to add either an image or '
+                         'an external link to an image.')
+            )
+
+        # certain cropping options do not work together, the following
+        # list defines the disallowed options used in the ``clean`` method
+        invalid_option_pairs = [
+            ('use_automatic_scaling', 'use_no_cropping'),
+            ('use_automatic_scaling', 'thumbnail_options'),
+            ('use_no_cropping', 'use_crop'),
+            ('use_no_cropping', 'use_upscale'),
+            ('use_no_cropping', 'thumbnail_options'),
+            ('thumbnail_options', 'use_crop'),
+            ('thumbnail_options', 'use_upscale'),
+        ]
+        # invalid_option_pairs
+        invalid_option_pair = None
+
+        for pair in invalid_option_pairs:
+            if getattr(self, pair[0]) and getattr(self, pair[1]):
+                invalid_option_pair = pair
+                break
+
+        if invalid_option_pair:
+            message = ugettext('The cropping selection is not valid. '
+                'You cannot combine "{field_a}" with "{field_b}".')
+            message = message.format(
+                field_a=self._meta.get_field(invalid_option_pair[0]).verbose_name,
+                field_b=self._meta.get_field(invalid_option_pair[1]).verbose_name,
+            )
+            raise ValidationError(message)