+++ tests/tests_models.py
@@ -1,15 +1,20 @@
 # -*- coding: utf-8 -*-
 from django.test import TestCase
-from django.utils.html import escapejs
+
 from djangocms_picture.models import Picture
+
+
+EXAMPLE_IMAGE = 'https://www.google.com/images/logo.png'
 
 
 class PictureTestCase(TestCase):
 
-    def test_escapejs_does_not_break_on_empty_picture(self):
-        """A Picture object can be passed to Django's escapejs"""
-        pic = Picture()
-        try:
-            escapejs(pic)
-        except TypeError:
-            self.fail("escapejs() raised TypeError on empty picture!")
+    def setUp(self):
+        Picture.objects.create(
+            external_picture=EXAMPLE_IMAGE,
+        )
+
+    def test_picture_instance(self):
+        """Picture instance has been created"""
+        picture = Picture.objects.get(external_picture=EXAMPLE_IMAGE)
+        self.assertEqual(picture.external_picture, EXAMPLE_IMAGE)