+++ CHANGELOG.rst
@@ -6,12 +6,14 @@ Changelog
 1.0.0 (unreleased)
 ==================
 
+* Added additional settings
 * Cleaned up file structure
-* Removed Django <1.8 support
+* Removed Django < 1.8 support
+* Removed ``alt`` attribute and migrated data to Filer
 * Adapted ``README.txt``
 
 
 0.2.0 (2015-02-12)
 ==================
 
-* Changes not tracked prior to 1.x
+* Use this version for Django < 1.8 support